Refactored code and implemented new features
The code from createLocaleManager.js was significantly refactored and the file was renamed to engine.js. We also added new search functionality in the server.js and its corresponding client side code in index.pug and search.pug. In addition, a new header was created with a search bar and drop-down menu. Styles for the header were added in styles.css. Lastly, the code for creating row elements was moved to its own mixin in mixins/row.pug and the implementation of the row was removed from index.pug.